project/requirements.txt
streamlit
pandas
numpy
requests
pyyaml
scikit-learn
plotly
matplotlib
altair
Pillow
openpyxl
statsmodels
